+++ CHANGELOG.md
@@ -10,7 +10,11 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Changed
 
+- [frontend]: The Logs in Home page has the same filter as the Logs page.
+
 ### Fixed
+
+- [frontend]: Fix Home for mobile.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">